Adjusted Game Adapter
When the user logs in and is not associated with a team, then the game
selection view is presented to the user. So that they can select a team
to follow.

It checks to see if the game id is null, if its null we present the game
selection view. If its not null, then we show that particular game.

Writing to the team selection to the local database is not complete.

https://gtg.atlassian.net/browse/AMA-1?atlOrigin=eyJpIjoiOWRmNTc2NWUwMmYwNDlmNWE5MzE2ZTZiMzRmMGQwNjAiLCJwIjoiaiJ9